add missing property names to docs/json
Prior to this patch, json schema validators would fail since
`properties` should be an `object` and not an `array`.

To fix that, this patch is replacing the `array` with an `object` and
providing proper keys to the internal children schemas.

Note: keys with an underline as a prefix are simple acronyms that match
the description of the schemas. They might not reflect the real property
names. The idea is to replace them with the proper values in the next
commits.