Cater for null callbacks array after destroy
After the destroy method has been called on `AnimationItem` the `_cbs`
array of event listeners is replaced with `null` which means that
subsequent calls to add, trigger or remove event listeners which tried
to access a particular name on that object resulted in "TypeError:
this._cbs is null". Checking that the `_cbs` array is truthy makes the
`addEventListener`, `triggerEvent`, and `removeEventListener` methods
safe to call in the same way as `AnimationItem.trigger` is even after
the animation has been destroyed.