Enable HImageElement to use preloaded images
The preloader previously used crossorigin='anonymous' when creating images,
while HImageElement did not. This resulted in those two requests recognized
as different by the browser, since one is a CORS request while the other
is not. In other words, the preloaded image was not used at all when
creating DOM image elements.